file renaming and auto-formatting
