+++ Util.cc
@@ -1,4 +1,4 @@
-#include "util.h"
+#include "Util.h"
 
 void log(uint32_t id, const string& logMessage) {
     time_t time = chrono::system_clock::to_time_t(chrono::system_clock::now());
